libbfddp: implement BFD protocol handling
Allow users to use the boiler plate BFD protocol handling code
(optionally) by including the `bfddp_extra.h` header.

The new library section implements the BFD session handling and calls
the appropriated callbacks for the hardware implementers to fill in with
their logic. This way we can have a BFD implementation that can be used
in any hardware and the only cost is implementing the hardware calls
from the library callbacks.